+++ tests/Unit/Validator/DateTimeValidatorTest.php
@@ -9,11 +9,12 @@
 
 declare(strict_types=1);
 
-namespace Eureka\Component\Validation\Tests\Validator;
+namespace Eureka\Component\Validation\Tests\Unit\Validator;
 
 use Eureka\Component\Validation\Exception\ValidationException;
 use Eureka\Component\Validation\Validator\DateTimeValidator;
 use Eureka\Component\Validation\ValidatorInterface;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -35,23 +36,23 @@ class DateTimeValidatorTest extends TestCase
      * @param  mixed $value
      * @param  mixed $excepted
      * @return void
-     * @dataProvider validValuesProvider
      */
-    public function testWithValidValues($value, $excepted): void
+    #[DataProvider('validValuesProvider')]
+    public function testWithValidValues(mixed $value, mixed $excepted): void
     {
-        $this->assertSame($excepted, $this->getValidator()->validate($value));
+        self::assertSame($excepted, $this->getValidator()->validate($value));
     }
 
     /**
      * @param  mixed $value
      * @param  mixed $excepted
      * @return void
-     * @dataProvider invalidValuesProvider
      */
-    public function testWithInvalidValues($value, $excepted): void
+    #[DataProvider('invalidValuesProvider')]
+    public function testWithInvalidValues(mixed $value, mixed $excepted): void
     {
         $this->expectException(ValidationException::class);
-        $this->assertSame($excepted, $this->getValidator()->validate($value));
+        self::assertSame($excepted, $this->getValidator()->validate($value));
     }
 
     /**
@@ -59,11 +60,11 @@ class DateTimeValidatorTest extends TestCase
      * @param  array<string,string|null|int|float|bool> $options
      * @param  mixed $excepted
      * @return void
-     * @dataProvider validValuesWithOptionsProvider
      */
-    public function testWithValidValueAndDefaultValues($value, array $options, $excepted): void
+    #[DataProvider('validValuesWithOptionsProvider')]
+    public function testWithValidValueAndDefaultValues(mixed $value, array $options, mixed $excepted): void
     {
-        $this->assertSame($excepted, $this->getValidator()->validate($value, $options));
+        self::assertSame($excepted, $this->getValidator()->validate($value, $options));
     }
 
     /**
@@ -71,17 +72,17 @@ class DateTimeValidatorTest extends TestCase
      * @param  array<string,string|null|int|float|bool> $options
      * @param  mixed $excepted
      * @return void
-     * @dataProvider invalidValuesWithOptionsProvider
      */
-    public function testWithInvalidValueAndDefaultValues($value, array $options, $excepted): void
+    #[DataProvider('invalidValuesWithOptionsProvider')]
+    public function testWithInvalidValueAndDefaultValues(mixed $value, array $options, mixed $excepted): void
     {
-        $this->assertSame($excepted, $this->getValidator()->validate($value, $options));
+        self::assertSame($excepted, $this->getValidator()->validate($value, $options));
     }
 
     /**
      * @return array<array<string|bool|int|float|null>>
      */
-    public function validValuesProvider(): array
+    public static function validValuesProvider(): array
     {
         return [
             ['2018-01-01 00:00:00', '2018-01-01 00:00:00'],
@@ -96,7 +97,7 @@ class DateTimeValidatorTest extends TestCase
     /**
      * @return array<array<string|bool|int|float|null>>
      */
-    public function invalidValuesProvider(): array
+    public static function invalidValuesProvider(): array
     {
         return [
             ['-2018-01-01 00:00:00', false],
@@ -108,7 +109,7 @@ class DateTimeValidatorTest extends TestCase
     /**
      * @return array<array<string|bool|int|float|null|array<string,string>>>
      */
-    public function validValuesWithOptionsProvider(): array
+    public static function validValuesWithOptionsProvider(): array
     {
         $default = '2001-01-01';
         $options = ['format' => 'Y-m-d H:i:s', 'format_output' => 'Y-m-d H:i:s', 'default' => $default];
@@ -118,16 +119,16 @@ class DateTimeValidatorTest extends TestCase
             ['2099-12-31 23:59:59', $options, '2099-12-31 23:59:59'],
             ['1901-07-01 12:00:00', $options, '1901-07-01 12:00:00'],
             ['2020-02-29 00:00:00', $options, '2020-02-29 00:00:00'],
-            ['12-31-2099 00:00:00', array_merge($options, ['format' => 'm-d-Y H:i:s', 'format_output' => 'd-m-Y H:i:s']), '31-12-2099 00:00:00'],
-            ['99-12-31 00:00:00',   array_merge($options, ['format' => 'y-m-d H:i:s']), '1999-12-31 00:00:00'],
-            ['01-07-1901 13:13',    array_merge($options, ['format' => 'd-m-Y H:i']), '1901-07-01 13:13:00'],
+            ['12-31-2099 00:00:00', \array_merge($options, ['format' => 'm-d-Y H:i:s', 'format_output' => 'd-m-Y H:i:s']), '31-12-2099 00:00:00'],
+            ['99-12-31 00:00:00',   \array_merge($options, ['format' => 'y-m-d H:i:s']), '1999-12-31 00:00:00'],
+            ['01-07-1901 13:13',    \array_merge($options, ['format' => 'd-m-Y H:i']), '1901-07-01 13:13:00'],
         ];
     }
 
     /**
      * @return array<array<string|bool|int|float|null|array<string,string>>>
      */
-    public function invalidValuesWithOptionsProvider(): array
+    public static function invalidValuesWithOptionsProvider(): array
     {
         $default = '2001-01-01 12:13:14';
         $options = ['format' => 'Y-m-d H:i:s', 'format_output' => 'Y-m-d H:i:s', 'default' => $default];