+++ tests/Unit/Validator/IntegerValidatorTest.php
@@ -9,11 +9,12 @@
 
 declare(strict_types=1);
 
-namespace Eureka\Component\Validation\Tests\Validator;
+namespace Eureka\Component\Validation\Tests\Unit\Validator;
 
 use Eureka\Component\Validation\Exception\ValidationException;
 use Eureka\Component\Validation\Validator\IntegerValidator;
 use Eureka\Component\Validation\ValidatorInterface;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -31,73 +32,51 @@ class IntegerValidatorTest extends TestCase
         return new IntegerValidator();
     }
 
-    /**
-     * @param  mixed $value
-     * @param  mixed $excepted
-     * @return void
-     * @dataProvider validValuesProvider
-     */
-    public function testWithValidValues($value, $excepted): void
+    #[DataProvider('validValuesProvider')]
+    public function testWithValidValues(mixed $value, mixed $excepted): void
     {
-        $this->assertSame($excepted, $this->getValidator()->validate($value));
+        self::assertSame($excepted, $this->getValidator()->validate($value));
     }
 
-    /**
-     * @param  mixed $value
-     * @param  mixed $excepted
-     * @return void
-     * @dataProvider invalidValuesProvider
-     */
-    public function testWithInvalidValues($value, $excepted): void
+    #[DataProvider('invalidValuesProvider')]
+    public function testWithInvalidValues(mixed $value, mixed $excepted): void
     {
         $this->expectException(ValidationException::class);
-        $this->assertSame($excepted, $this->getValidator()->validate($value));
+        self::assertSame($excepted, $this->getValidator()->validate($value));
     }
 
     /**
-     * @param  mixed $value
      * @param  array<string,string|null|int|float|bool> $options
-     * @param  mixed $excepted
-     * @return void
-     * @dataProvider validValuesWithOptionsProvider
      */
-    public function testWithValidValueAndDefaultValues($value, array $options, $excepted): void
+    #[DataProvider('validValuesWithOptionsProvider')]
+    public function testWithValidValueAndDefaultValues(mixed $value, array $options, mixed $excepted): void
     {
-        $this->assertSame($excepted, $this->getValidator()->validate($value, $options));
+        self::assertSame($excepted, $this->getValidator()->validate($value, $options));
     }
 
     /**
-     * @param  mixed $value
      * @param  array<string,string|null|int|float|bool> $options
-     * @param  mixed $excepted
-     * @return void
-     * @dataProvider invalidValuesWithOptionsProvider
      */
-    public function testWithInvalidValueAndDefaultValues($value, array $options, $excepted): void
+    #[DataProvider('invalidValuesWithOptionsProvider')]
+    public function testWithInvalidValueAndDefaultValues(mixed $value, array $options, mixed $excepted): void
     {
-        $this->assertSame($excepted, $this->getValidator()->validate($value, $options));
+        self::assertSame($excepted, $this->getValidator()->validate($value, $options));
     }
 
-    /**
-     * @return void
-     */
     public function testWithOctalValueAsValidValues(): void
     {
-        $this->assertSame(octdec('777'), $this->getValidator()->validate(0777, [], FILTER_FLAG_ALLOW_OCTAL));
+        self::assertSame(octdec('777'), $this->getValidator()->validate(0777, [], FILTER_FLAG_ALLOW_OCTAL));
     }
 
-    /**
-     * @return void
-     */
     public function testWithHexadecimalValueAsValidValues(): void
     {
-        $this->assertSame(hexdec('fa11'), $this->getValidator()->validate('0xfa11', [], FILTER_FLAG_ALLOW_HEX));
+        self::assertSame(hexdec('fa11'), $this->getValidator()->validate('0xfa11', [], FILTER_FLAG_ALLOW_HEX));
     }
 
     /**
      * @return array<array<string|bool|int|float|null>>
      */
-    public function validValuesProvider(): array
+    public static function validValuesProvider(): array
     {
         return [
             [0.0,  0],
@@ -110,7 +89,7 @@ class IntegerValidatorTest extends TestCase
     /**
      * @return array<array<string|bool|int|float|null>>
      */
-    public function invalidValuesProvider(): array
+    public static function invalidValuesProvider(): array
     {
         return [
             [1.20, 1],
@@ -128,7 +107,7 @@ class IntegerValidatorTest extends TestCase
     /**
      * @return array<array<string|bool|int|float|null|array<string,string|null|int|float|bool>>>
      */
-    public function validValuesWithOptionsProvider(): array
+    public static function validValuesWithOptionsProvider(): array
     {
         $default = -1;
         $options = ['default' => $default];
@@ -137,16 +116,16 @@ class IntegerValidatorTest extends TestCase
             [0.0, ['default' => $default], 0],
             [15, ['default' => $default, 'min_range' => 0, 'max_range' => 100], 15],
             ['42', ['default' => $default, 'min_range' => 42, 'max_range' => 42], 42],
-            [-128,  array_merge($options, IntegerValidator::TINYINT_SIGNED), -128],
-            [127,  array_merge($options, IntegerValidator::TINYINT_SIGNED), 127],
-            [128,  array_merge($options, IntegerValidator::TINYINT_UNSIGNED), 128],
+            [-128, \array_merge($options, IntegerValidator::TINYINT_SIGNED), -128],
+            [127,  \array_merge($options, IntegerValidator::TINYINT_SIGNED), 127],
+            [128,  \array_merge($options, IntegerValidator::TINYINT_UNSIGNED), 128],
         ];
     }
 
     /**
      * @return array<array<string|bool|int|float|null|array<string,string|null|int|float|bool>>>
      */
-    public function invalidValuesWithOptionsProvider(): array
+    public static function invalidValuesWithOptionsProvider(): array
     {
         $default = -1;
         $options = ['default' => $default];
@@ -156,9 +135,9 @@ class IntegerValidatorTest extends TestCase
             [null, $options, $default],
             [1.20, ['default' => $default], $default],
             [1.20, ['default' => $default, 'min_range' => 1, 'max_range' => 2], $default],
-            [1.01, array_merge($options, ['min_range' => 0, 'max_range' => 1]), $default],
-            [254,  array_merge($options, IntegerValidator::TINYINT_SIGNED), $default],
-            [256,  array_merge($options, IntegerValidator::TINYINT_UNSIGNED), $default],
+            [1.01, \array_merge($options, ['min_range' => 0, 'max_range' => 1]), $default],
+            [254,  \array_merge($options, IntegerValidator::TINYINT_SIGNED), $default],
+            [256,  \array_merge($options, IntegerValidator::TINYINT_UNSIGNED), $default],
         ];
     }
 }